fix vote return value dictionary keys
`vote` function previously used `self.players` as the key list in
`dict(zip(...))`. This means that for a vote that doesn't include all
players (e.g. quest betrayal vote), the keys would be wrong. The keys
are never inspected in the partial-vote case, but the function should be
fixed regardless.

Signed-off-by: Tal Shorer <tal.shorer@gmail.com>